--- a/pkg/gossip/gossip_test.go
+++ b/pkg/gossip/gossip_test.go
@@ -1058,3 +1058,70 @@ func TestServerSendsHighStampsDiff(t *testing.T) {
 		return nil
 	})
 }
+
+// TestCallbacksPendingMetricGoesToZeroOnStop verifies that the CallbacksPending
+// metric is correctly decremented when a callback is unregistered with pending work
+// or when the stopper is stopped.
+func TestCallbacksPendingMetricGoesToZeroOnStop(t *testing.T) {
+	defer leaktest.AfterTest(t)()
+
+	testCases := []struct {
+		name    string
+		cleanup func(g *Gossip, unregister func(), stopper *stop.Stopper, ctx context.Context)
+	}{
+		{
+			name: "unregister callback",
+			cleanup: func(g *Gossip, unregister func(), stopper *stop.Stopper, ctx context.Context) {
+				unregister()
+			},
+		},
+		{
+			name: "stopper shutdown",
+			cleanup: func(g *Gossip, unregister func(), stopper *stop.Stopper, ctx context.Context) {
+				stopper.Stop(ctx)
+			},
+		},
+	}
+
+	for _, tc := range testCases {
+		t.Run(tc.name, func(t *testing.T) {
+			ctx := context.Background()
+			stopper := stop.NewStopper()
+			defer stopper.Stop(ctx)
+			g := NewTest(1, stopper, metric.NewRegistry())
+
+			unregister := g.RegisterCallback("test.*", func(key string, val roachpb.Value) {
+				// Do nothing.
+			})
+
+			// Add 100 infos to the gossip that will be processed by the callback.
+			for i := 0; i < 100; i++ {
+				slice := []byte("b1")
+				require.NoError(t, g.AddInfo(fmt.Sprintf("test.key%d", i), slice, time.Hour))
+			}
+
+			// Execute the cleanup action (either unregister or stopper.Stop)
+			// We do this in a goroutine to help cause interesting potential race conditions.
+			go func() {
+				tc.cleanup(g, unregister, stopper, ctx)
+			}()
+
+			// Add another 100 infos to the gossip that will be processed by the callback.
+			// We do this in a goroutine to help cause interesting potential race conditions.
+			go func() {
+				for i := 0; i < 100; i++ {
+					slice := []byte("b2")
+					require.NoError(t, g.AddInfo(fmt.Sprintf("test.key%d", i), slice, time.Hour))
+				}
+			}()
+
+			// Wait for the pending callbacks metric to go to 0.
+			testutils.SucceedsSoon(t, func() error {
+				if g.mu.is.metrics.CallbacksPending.Value() != 0 {
+					return fmt.Errorf("CallbacksPending should be 0, got %d", g.mu.is.metrics.CallbacksPending.Value())
+				}
+				return nil
+			})
+		})
+	}
+}